website/requirements.txt
asgiref==3.8.1
Django==4.2
pillow==10.4.0
sqlparse==0.5.1
psycopg2==2.9.9
djangorestframework==3.15.2
channels==4.1.0
channels_redis==4.2.0
django-cors-headers
uvicorn
msgpack==1.0.8
redis==5.0.8
